add negative (, and power support

+++ Mab.StringLibrary/Formula/PostfixConverter.cs
@@ -17,7 +17,7 @@ namespace Mab.StringLibrary.Formula
 		}
 
 
-		public decimal ProcessPostfix()
+		public double ProcessPostfix()
 		{
             try
             {
